requirements.txt
numpy~=1.20.1
streamlit~=1.8.1
torch~=1.7.1
pillow~=8.1.2
torchray~=1.0.0.2
matplotlib~=3.3.4
scikit-image~=0.19.2
git+https://github.com/openai/CLIP.git
